Reduce code duplication in entities
