+++ custom_components/google_home/entity.py
@@ -1,27 +1,13 @@
 """Defines base entities for Google Home"""
 
-from typing import Set, Tuple, TypedDict
+from typing import Any, Dict, List, Set, Tuple, TypedDict
 
-from homeassistant.const import DEVICE_CLASS_TIMESTAMP
 from homeassistant.helpers.update_coordinator import (
     CoordinatorEntity,
     DataUpdateCoordinator,
 )
 
-from .const import (
-    DEFAULT_NAME,
-    DOMAIN,
-    ICON_ALARMS,
-    ICON_TIMERS,
-    ICON_TOKEN,
-    LABEL_ALARMS,
-    LABEL_DEVICE,
-    LABEL_NEXT_ALARM,
-    LABEL_NEXT_TIMER,
-    LABEL_TIMERS,
-    MANUFACTURER,
-    VERSION,
-)
+from .const import DEFAULT_NAME, DOMAIN, MANUFACTURER
 
 
 class DeviceInfo(TypedDict):
@@ -29,12 +15,11 @@ class DeviceInfo(TypedDict):
 
     identifiers: Set[Tuple[str, str]]
     name: str
-    model: str
     manufacturer: str
 
 
-class GoogleHomeDeviceEntity(CoordinatorEntity):
-    """Entity base for device sensor"""
+class GoogleHomeBaseEntity(CoordinatorEntity):
+    """Base entity base for Google Home sensors"""
 
     def __init__(self, coordinator: DataUpdateCoordinator, device_name: str):
         super().__init__(coordinator)
@@ -45,164 +30,22 @@ class GoogleHomeDeviceEntity(CoordinatorEntity):
         return {
             "identifiers": {(DOMAIN, self.device_name)},
             "name": f"{DEFAULT_NAME} {self.device_name}",
-            "model": VERSION,
             "manufacturer": MANUFACTURER,
         }
 
-    @property
-    def name(self) -> str:
-        """Return the name of the sensor."""
-        return f"{self.device_name} {LABEL_DEVICE}"
+    def get_device(self):
+        """Return the device matched by device name
+        from the list of google devices in coordinator_data"""
+        return next(
+            (
+                device
+                for device in self.coordinator.data
+                if device.name == self.device_name
+            ),
+            None,
+        )
 
-    @property
-    def unique_id(self) -> str:
-        """Return a unique ID to use for this entity."""
-        return f"{self.device_name}/{LABEL_DEVICE}"
-
-    @property
-    def icon(self) -> str:
-        """Return the icon of the sensor."""
-        return ICON_TOKEN
-
-    @property
-    def device_class(self) -> str:
-        """Return the device class of the sensor."""
-        return "google_home__custom_device_class"
-
-
-class GoogleHomeAlarmEntity(CoordinatorEntity):
-    """Entity base for Alarm sensor"""
-
-    def __init__(self, coordinator: DataUpdateCoordinator, device_name: str):
-        super().__init__(coordinator)
-        self.device_name = device_name
-
-    @property
-    def device_info(self) -> DeviceInfo:
-        return {
-            "identifiers": {(DOMAIN, self.device_name)},
-            "name": f"{DEFAULT_NAME} {self.device_name}",
-            "model": VERSION,
-            "manufacturer": MANUFACTURER,
-        }
-
-    @property
-    def name(self) -> str:
-        """Return the name of the sensor."""
-        return f"{self.device_name} {LABEL_ALARMS}"
-
-    @property
-    def unique_id(self) -> str:
-        """Return a unique ID to use for this entity."""
-        return f"{self.device_name}/{LABEL_ALARMS}"
-
-    @property
-    def icon(self) -> str:
-        """Icon to use in the frontend, if any."""
-        return ICON_ALARMS
-
-
-class GoogleHomeNextAlarmEntity(CoordinatorEntity):
-    """Entity base for next alarm sensor"""
-
-    def __init__(self, coordinator: DataUpdateCoordinator, device_name: str):
-        super().__init__(coordinator)
-        self.device_name = device_name
-
-    @property
-    def device_info(self) -> DeviceInfo:
-        return {
-            "identifiers": {(DOMAIN, self.device_name)},
-            "name": f"{DEFAULT_NAME} {self.device_name}",
-            "model": VERSION,
-            "manufacturer": MANUFACTURER,
-        }
-
-    @property
-    def name(self) -> str:
-        """Return the name of the sensor."""
-        return f"{self.device_name} {LABEL_NEXT_ALARM}"
-
-    @property
-    def unique_id(self) -> str:
-        """Return a unique ID to use for this entity."""
-        return f"{self.device_name}/{LABEL_NEXT_ALARM}"
-
-    @property
-    def icon(self) -> str:
-        """Icon to use in the frontend, if any."""
-        return ICON_ALARMS
-
-    @property
-    def device_class(self) -> str:
-        """Return the device class of the sensor."""
-        return DEVICE_CLASS_TIMESTAMP
-
-
-class GoogleHomeTimersEntity(CoordinatorEntity):
-    """Entity base for timers sensor"""
-
-    def __init__(self, coordinator: DataUpdateCoordinator, device_name: str):
-        super().__init__(coordinator)
-        self.device_name = device_name
-
-    @property
-    def name(self) -> str:
-        """Return the name of the sensor."""
-        return f"{self.device_name} {LABEL_TIMERS}"
-
-    @property
-    def device_info(self) -> DeviceInfo:
-        return {
-            "identifiers": {(DOMAIN, self.device_name)},
-            "name": f"{DEFAULT_NAME} {self.device_name}",
-            "model": VERSION,
-            "manufacturer": MANUFACTURER,
-        }
-
-    @property
-    def unique_id(self) -> str:
-        """Return a unique ID to use for this entity."""
-        return f"{self.device_name}/{LABEL_TIMERS}"
-
-    @property
-    def icon(self) -> str:
-        """Icon to use in the frontend, if any."""
-        return ICON_TIMERS
-
-
-class GoogleHomeNextTimerEntity(CoordinatorEntity):
-    """Entity base for next timer sensor"""
-
-    def __init__(self, coordinator: DataUpdateCoordinator, device_name: str):
-        super().__init__(coordinator)
-        self.device_name = device_name
-
-    @property
-    def device_info(self) -> DeviceInfo:
-        return {
-            "identifiers": {(DOMAIN, self.device_name)},
-            "name": f"{DEFAULT_NAME} {self.device_name}",
-            "model": VERSION,
-            "manufacturer": MANUFACTURER,
-        }
-
-    @property
-    def name(self) -> str:
-        """Return the name of the sensor."""
-        return f"{self.device_name} {LABEL_NEXT_TIMER}"
-
-    @property
-    def unique_id(self) -> str:
-        """Return a unique ID to use for this entity."""
-        return f"{self.device_name}/{LABEL_NEXT_TIMER}"
-
-    @property
-    def icon(self) -> str:
-        """Icon to use in the frontend, if any."""
-        return ICON_TIMERS
-
-    @property
-    def device_class(self) -> str:
-        """Return the device class of the sensor."""
-        return DEVICE_CLASS_TIMESTAMP
+    @staticmethod
+    def as_dict(obj_list: List[Any]) -> List[Dict[Any, Any]]:
+        """Return list of objects represented as dictionaries """
+        return [obj.__dict__ for obj in obj_list]
